// web/src/shared/api/generated/RuntimeLogEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RuntimeLogLevel } from "./RuntimeLogLevel";

export type RuntimeLogEntry = { id: string, timestamp: string, level: RuntimeLogLevel, module: string, target: string, message: string, summary: string, fields: { [key in string]: string }, };